% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fROH.r
\name{fROH}
\alias{fROH}
\title{Estimating consanguinity levels from runs of homozygosity}
\usage{
fROH(
  x,
  dist.units = c("bases", "cM"),
  NAs = c("het", "hom"),
  minNbSNPs = 100L,
  minROHlength,
  minDistHet,
  maxGapLength,
  beg = 1L,
  end = ncol(x)
)
}
\arguments{
\item{x}{a bed.matrix}

\item{dist.units}{distance unit to use (either 'bases' or 'cM')}

\item{NAs}{Should missing data be treated as heterozygous or homozygous?}

\item{minNbSNPs}{minimal number of SNPS in a ROH}

\item{minROHlength}{minimal length of a ROH (in dist.units)}

\item{minDistHet}{minimal distance between hetorozygous SNPs (in dist.units)}

\item{maxGapLength}{maximal distance between two SNPs in a ROH (in dist.units=}

\item{beg}{index of the first SNP to consider}

\item{end}{index of the last SNP to consider}
}
\value{
A data frame with a line for each individual in `x`, and columns `nbSNPs` 
(the number SNPs in ROHs), `nbSegments` (the number of ROH segments)
`length` (the total length of ROH segments) and `fROH` (the estimated coefficient of
consanguinity).
}
\description{
Estimates consanguinity levels as the proportion of the genome in runs of homozygosity
}
\details{
See `\link{ROH}` for details on the definition of runs of hmozygosity
and for a detailed description of the parameters.
This functions estimates the coefficient of consanguinity \eqn{f} of each individual
in `x` as the proportion of its genome in runs of homozygosity.
}
\seealso{
`\link{ROH}`
}
