% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-bed.matrix-methods.r,
%   R/class-bed.matrix-product.r, R/class-bed.matrix-subsetting.r
\name{bed.matrix-methods}
\alias{bed.matrix-methods}
\alias{\%*\%,bed.matrix,matrix-method}
\alias{\%*\%,matrix,bed.matrix-method}
\alias{\%*\%,bed.matrix,vector-method}
\alias{\%*\%,vector,bed.matrix-method}
\alias{[,bed.matrix,numeric,missing,missing-method}
\alias{[,bed.matrix,logical,missing,missing-method}
\alias{[,bed.matrix,missing,numeric,missing-method}
\alias{[,bed.matrix,missing,logical,missing-method}
\alias{[,bed.matrix,logical,logical,missing-method}
\alias{[,bed.matrix,logical,numeric,missing-method}
\alias{[,bed.matrix,numeric,logical,missing-method}
\alias{[,bed.matrix,numeric,numeric,missing-method}
\title{Methods for the bed.matrix class}
\usage{
\S4method{\%*\%}{bed.matrix,matrix}(x, y)

\S4method{\%*\%}{matrix,bed.matrix}(x, y)

\S4method{\%*\%}{bed.matrix,vector}(x, y)

\S4method{\%*\%}{vector,bed.matrix}(x, y)

\S4method{[}{bed.matrix,numeric,missing,missing}(x, i, j)

\S4method{[}{bed.matrix,logical,missing,missing}(x, i, j)

\S4method{[}{bed.matrix,missing,numeric,missing}(x, i, j)

\S4method{[}{bed.matrix,missing,logical,missing}(x, i, j)

\S4method{[}{bed.matrix,logical,logical,missing}(x, i, j)

\S4method{[}{bed.matrix,logical,numeric,missing}(x, i, j)

\S4method{[}{bed.matrix,numeric,logical,missing}(x, i, j)

\S4method{[}{bed.matrix,numeric,numeric,missing}(x, i, j)
}
\arguments{
\item{x, y}{a bed matrix, or a R matrix, depending on the method}

\item{i, j}{numeric or logical vectors of indices}
}
\description{
Methods for the bed.matrix class
}
\details{
"%*%" allow to multiply a bed matrix by a numeric vector, to the left or to the right.
Depending of the standardization of the matrix, the genotypes will be interpreted as 0, 1 or 2,
or as standardized values.

The bracket operator "[" allow to extract sub matrices pretty much as for R matrices.
}
