% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROH.r
\name{ROH}
\alias{ROH}
\title{Runs of Homozygosity}
\usage{
ROH(
  x,
  dist.units = c("bases", "cM"),
  NAs = c("het", "hom"),
  minNbSNPs = 100L,
  minROHlength,
  minDistHet,
  maxGapLength,
  beg = 1L,
  end = ncol(x)
)
}
\arguments{
\item{x}{a bed.matrix}

\item{dist.units}{distance unit to use (either 'bases' or 'cM')}

\item{NAs}{Should missing data be treated as heterozygous or homozygous?}

\item{minNbSNPs}{minimal number of SNPS in a ROH}

\item{minROHlength}{minimal length of a ROH (in dist.units)}

\item{minDistHet}{minimal distance between hetorozygous SNPs (in dist.units)}

\item{maxGapLength}{maximal distance between two SNPs in a ROH (in dist.units=}

\item{beg}{index of the first SNP to consider}

\item{end}{index of the last SNP to consider}
}
\value{
A list of data frames, one for each indivual in `x`, describing the ROHs.
These data frames have columns `i.beg`, the index of the first SNPs in the ROH,
`i.end`, the index of the last SNP in the ROH, `pos.beg` and `pos.end` their genomic
position (in accordance with `dist.units`) and `nb.het`, the number of heterozygous
SNPs in the ROH.
}
\description{
Identifies Runs of Homozygosity
}
\details{
Runs of homozygosity (ROHs) are sequences of homozygous SNPs, including a minimal number 
`minNbSNPs` of SNPs, and spanning a minimal length of `minROHlength`, in units specified
by `dist.units`. They can contain some heterozygous SNPs, if these are at distance 
larger than `minDistHet` from the extremities of the ROH, and from each other.
Consecutive SNPs in a ROH are at a distance smaller than `maxGepLength`.

The distance units are either base pairs ("bases") or centiMorgans ("cM").
It is better to use `dist.units = "cM"` but this requires that the `dist`
column in `x@snps` is set (see `\link{set.dist}`).

If `minROHlength`, `minDistHet` and `maxGapLength` are not specified, default values
are set depending on the units chosen as follows: if `dist.units = "bases"` 
`minROHlength = 1e6` (1Mb), `minDistHet = 5e5` (500 kb) and `maxGapLength = 1e6`.
If `dist.units = "cM"`, `minROHlength = 1`, `minDistHet = 0.5` and `maxGapLength = 1`.
}
\seealso{
`\link{set.dist}`, `\link{fROH}`
}
